test-spexs-python/pyproject.toml
[project]
name = "test-spexs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "hubspot-api-client>=12.0.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "pandas>=2.0.0",
    "pymongo>=4.15.1",
]
